pyproject.toml
[project]
name = "cog-generator"
version = "0.1.0"
description = "Utils to generate COG files using i/o from S3."
authors = [
    {name = "Sam Woodcock", email = "samuel.woodcock@wsl.ch"},
]
dependencies = [
    "envidat-utils[dotenv]>=0.5.1",
    "rio-cogeo>=3.2.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}
keywords = ["envidat", "cog", "geotiff", "s3"]

[tool.pdm]
[tool.pdm.dev-dependencies]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
